Reimplemented dronekit to decouple from sensors
