+++ src/asv_loyola_us/asv_loyola_us/dronekit_node.py
@@ -3,119 +3,171 @@ from rclpy.node import Node
 from rclpy.action import ActionServer, CancelResponse, GoalResponse, ActionClient
 from rclpy.executors import MultiThreadedExecutor
 from rclpy.callback_groups import ReentrantCallbackGroup
-#TODO: deprecate dronekit, use pymavlink
+
 from dronekit import connect, VehicleMode, LocationGlobal
 import numpy as np
 import pymavlink
 import traceback
 from math import atan2
-from .submodulos.dictionary import dictionary
 import time
 from pymavlink.dialects.v20 import ardupilotmega as mavlink2 #for obstacle distance information
 from numpy import uint
-#import intefaces
+
 from asv_interfaces.msg import Status, Nodeupdate, Camera, Obstacles, Location
 from asv_interfaces.srv import CommandBool, ASVmode, Newpoint, Takesample
-from asv_interfaces.action import Goto, SensorSample
+from asv_interfaces.action import Goto
 
 
 import os
-os.environ["MAVLINK20"] = "1"
+os.environ['MAVLINK20'] = '1'
 
-# This node generates the necessary services for  comunication towards the drone
-#parameters are only read at start
+
+ardupilotModeDictionary = {
+
+    "ASV_MODE": {
+        "0":  "MANUAL",
+        "1":  "CIRCLE",
+        "2":  "STABILIZE",
+        "3":  "TRAINING",
+        "4":  "ACRO",
+        "5":  "FBWA",
+        "6":  "FBWB",
+        "7":  "CRUISE",
+        "8":  "AUTOTUNE",
+        "10": "AUTO",
+        "11": "RTL",
+        "12": "LOITER",
+        "13": "TAKEOFF",
+        "14": "AVOID_ADSB",
+        "15": "GUIDED",
+        "16": "INITIALISING",
+        "17": "QSTABILIZE",
+        "18": "QHOVER",
+        "19": "QLOITER",
+        "20": "QLAND",
+        "21": "QRTL",
+        "22": "QAUTOTUNE",
+        "23": "QACRO",
+        "24": "THERMAL"
+        },
+    "MISSION_MODE": {
+        "0": "STANDBY",
+        "1": "GUIDED",
+        "2": "MANUAL",
+        "3": "SIMPLE",
+        "4": "RTL"
+        }
+
+    }
+
 
 class Dronekit_node(Node):
 
+    """ This node is in charge of the communication with the drone using dronekit interface.
+    It implements the following services/actions """
 
-    #TODO define a wildcard_caldback for dronekit criticals
-    """
-       def wildcard_callback(self, attr_name, value):
-           self.get_logger().fatal(f"dronekit error {attr_name}\n value: {value}")
-       self.vehicle.add_attribute_listener('*', wildcard_callback)
-    """
+    # Services:
+    #   - CommandBool: to arm/disarm the drone
+    #   - ASVmode: to change the mode of the drone
+    #   - Resethome: to reset the home location of the drone
+
+    # Actions:
+    #   - Goto: to send the drone to a given location/s
 
 
-    #his functions defines and assigns value to the parameters
     def parameters(self):
 
+        # Declare the parameters of the node
+
+        # Parameter to enable/disable the debug mode
         self.declare_parameter('debug', False)
         self.DEBUG = self.get_parameter('debug').get_parameter_value().bool_value
 
-
+        # Parameter to set the vehicle IP
         self.declare_parameter('vehicle_ip', 'tcp:navio.local:5678')
         self.declare_parameter('debug_vehicle_ip', 'tcp:127.0.0.1:5760')
-
-        if self.DEBUG:
+        # If debug mode is enabled, the vehicle IP is set to the debug_vehicle_ip parameter
+        if not self.DEBUG:
             self.vehicle_ip = self.get_parameter('vehicle_ip').get_parameter_value().string_value
         else:
             self.vehicle_ip = self.get_parameter('debug_vehicle_ip').get_parameter_value().string_value
 
+        # Parameter to set the connection timeout
         self.declare_parameter('connect_timeout', 15)
         self.connect_timeout = self.get_parameter('connect_timeout').get_parameter_value().integer_value
+        # Parameter to set the maximum distance of a Waypoint
         self.declare_parameter('max_distance', 5000)
         self.max_distance = self.get_parameter('max_distance').get_parameter_value().integer_value
+        # The vehicle ID #
         self.declare_parameter('vehicle_id', 1)
         self.vehicle_id=self.get_parameter('vehicle_id').get_parameter_value().integer_value
+        # The vehicle arm timeout in seconds 
         self.declare_parameter('arm_timeout', 15)
         self.arm_timeout = self.get_parameter('arm_timeout').get_parameter_value().integer_value
+        # The vehicle travellin timeout in seconds
         self.declare_parameter('travelling_timeout', 60)
         self.travelling_timeout = self.get_parameter('travelling_timeout').get_parameter_value().integer_value
-        
-        
+        # WP_RADIUS
+        self.declare_parameter('wp_radius', 0.5)
+        self.wp_radius = self.get_parameter('wp_radius').get_parameter_value().double_value
 
-
-    #this function declares the services, its only purpose is to keep code clean
     def declare_services(self):
-        #host
+        """ Declare the services of the node: i) the host services, ii) the client services """
+
+        # i) Host services
         self.arm_vehicle_service = self.create_service(CommandBool, 'arm_vehicle', self.arm_vehicle_callback)
-        self.change_ASV_mode_service = self.create_service(ASVmode, 'change_asv_mode', self.change_asv_mode_callback)
+        self.mode_vehicle_service = self.create_service(ASVmode, 'change_asv_mode', self.change_asv_mode_callback)
         self.reset_home_service = self.create_service(CommandBool, 'reset_home', self.reset_home_callback)
 
-        #client
-        self.asv_mission_mode_client = self.create_client(ASVmode, 'change_mission_mode')
-        self.take_sample_client = self.create_client(Takesample, 'get_sample')
+        # ii) Client services
         self.calculate_path_client = self.create_client(Newpoint, 'calculate_path')
+
+    def declare_actions(self):
+        """ Declare the actions of the node """
+
+        # GoTo action #
+        self.go_to_server = ActionServer(self, Goto, 'goto', execute_callback=self.goto_execute_callback,
+                                         goal_callback=self.goto_check_goal_callback,
+                                         cancel_callback=self.goto_cancel_callback,
+                                         callback_group=ReentrantCallbackGroup())
 
 
     def declare_topics(self):
-        timer_period = 0.5  # seconds
+        """ Declare the topics of the node """
+
+        timer_period = 0.1  # seconds
         self.status_publisher = self.create_publisher(Status, 'status', 10)
         self.status_publisher_timer = self.create_timer(timer_period, self.status_publish)
-        self.obstacles_subscriber = self.create_subscription(Obstacles, 'camera_obstacles', self.camera_obstacles_callback, 10)
         self.waypoints_publisher = self.create_publisher(Location, 'waypoint_mark', 10)
 
-
-    def declare_actions(self):
-        self.go_to_server = ActionServer(self, Goto, 'goto', execute_callback=self.goto_execute_callback,
-                                         goal_callback=self.goto_accept,
-                                         handle_accepted_callback=self.goto_accepted_callback,
-                                         cancel_callback=self.goto_cancel,
-                                         callback_group=ReentrantCallbackGroup())
-        self.goto_goal_handle = None
-        #TODO: if we want to call this server more than once at a time consider using multithread executor in sping and reentrant_callback_group in action servers
-        self.sensor_action_client = ActionClient(self, SensorSample, 'sensor_read') #action client
-
+    
     def __init__(self):
-        # start the node
-        super().__init__('Dronekit_node')
+        """ Constructor """
 
-        # declare parameters
+        super().__init__('dronekit_node')
+
+        # Declare the parameters of the node
         self.parameters()
+        # Initialize the status
         self.status = Status()
-
-        self.last_obstacle_distance_sent_ms=0
         # connect to vehicle
+
         if self.DEBUG:
             self.get_logger().warning("Debug mode enabled")
+
         self.get_logger().info(f"Connecting to vehicle in {self.vehicle_ip}")
+
         try:
+            # Connect to the vehicle using the vehicle IP
             self.vehicle = connect(self.vehicle_ip, timeout=self.connect_timeout, source_system=1, source_component=93)
-            self.dictionary()
+            self.get_logger().info(f"Connected to vehicle in {self.vehicle_ip}")
+            # Create the timer to read the RC channels
             self.rc_read_timer=self.create_timer(0.1, self.rc_read_callback)
+            # Declare the services and other stuff
             self.declare_topics()
             self.declare_services()
             self.declare_actions()
+            # Initialize the vehicle
             self.vehicle.add_message_listener('SYS_STATUS', self.vehicle_status_callback)
             self.cmds = self.vehicle.commands
             self.get_logger().info(f"Waiting for initialization")
@@ -123,52 +175,100 @@ class Dronekit_node(Node):
             self.cmds.wait_ready()
             self.home = self.vehicle.home_location
             self.get_logger().info(f"Vehicle home location is {self.home}")
+
         except ConnectionRefusedError:
             self.get_logger().error(f"Connection to navio2 was refused")
             self.get_logger().fatal("Drone module is dead")
             self.destroy_node()
+
         except OSError:
             self.get_logger().error(f"Navio2 was not found in the same network")
             self.get_logger().fatal("Drone module is dead")
             self.destroy_node()
+
         except TimeoutError:
             self.get_logger().error(f"Navio2 port was busy, timeout error")
             self.get_logger().fatal("Drone module is dead")
             self.destroy_node()
+
         except:
             error = traceback.format_exc()
             self.get_logger().error(f"Connection to navio2 could not be made, unknown error:\n {error}")
             self.get_logger().fatal("Drone module is dead")
             self.destroy_node()
 
-        
+    def vehicle_status_callback(self, vehicle, name, msg):
+        pass
+
+    def reset_home_callback(self, request, response):
+        self.vehicle.home_location = self.vehicle.location.global_frame
 
     def arm_vehicle_callback(self, request, response):
+        """ Vehicle arm service callback:
+
+        request.value: False to disarm, True to arm
+
         """
-        Arming vehicle function.
-        Args:
-             request.value:
-                -True : arm vehicle
-                -False: disarm vehicle
-        """
+
+        # First, read the current status of the vehicle
         if self.status.manual_mode:
-            self.get_logger().info(f"RC in manual, cannot arm or disarm")
+            # In manual mode, the vehicle cannot be armed because 
+            # it is not safe to override the RC
+            self.get_logger().error("RC IN MANUAL -> Vehicle cannot be armed/disarmed in manual mode")
+            response.success = False
             return response
-        try:
-            if request.value:
-                self.vehicle.arm()
-                self.get_logger().info('Vehicle armed')
-                response.success=True
-            else:
-                self.vehicle.disarm()
-                self.get_logger().info('Vehicle disarmed')
-                response.success=True
-        except:
-            self.get_logger().error('Couldn\'t arm vehicle')
-            response.success=False
+        else:
+
+            try:
+                if request.value:
+                    self.vehicle.arm()
+                    self.get_logger().info("Vehicle armed")
+                else:
+                    self.vehicle.disarm()
+                    self.get_logger().info("Vehicle disarmed")
+                # Transmit the new status
+                while self.vehicle.armed != request.value:
+                    self.get_logger().info("Waiting for vehicle to arm/disarm")
+                    time.sleep(0.1)
+
+            except:
+                self.get_logger().error("Vehicle could not be armed/disarmed")
+                response.success = False
+        
         return response
 
+    def rc_read_callback(self):
+        if self.DEBUG==False:
+            #if there is no RC return home
+            try:
+                #manage no RC detected
+                if all([self.vehicle.channels['5'] == 0, self.vehicle.channels['6'] == 0]):
+                    if self.vehicle.mode != VehicleMode("RTL"):
+                        self.get_logger().info("seems there is no RC connected", once=True)
+                        self.status.manual_mode = False #override RC if there is no connection
+                        
+                #manage RC switch between auto and manual mode
+                elif self.status.manual_mode!=(self.vehicle.channels['6']>1200):
+                    self.get_logger().info("manual interruption" if self.vehicle.channels['6']>1200 else "automatic control regained")
+                    self.status.manual_mode=bool(self.vehicle.channels['6']>1200)
+
+                #manage arm when RC in manual
+                elif self.status.manual_mode:                
+                    if self.vehicle.mode != VehicleMode("MANUAL") : #vehicle is not in desired mode
+                        self.get_logger().info("manual switching vehicle mode to manual")
+                        self.vehicle.mode = VehicleMode("MANUAL")
+                    if self.vehicle.armed!=(self.vehicle.channels['5']>1200):
+                        if self.vehicle.channels['5']>1200:
+                            self.vehicle.arm()
+                            self.get_logger().info("manual arm")
+                        else:
+                            self.vehicle.disarm()
+                            self.get_logger().info("manual disarm")
+            except:
+                pass
+
     def status_publish(self):
+
         try:
             # TODO: es necesario que dronekit te devuelva valores no corruptos, crash otherwise
             self.status.lat = self.vehicle.location.global_relative_frame.lat
@@ -176,11 +276,13 @@ class Dronekit_node(Node):
             self.status.yaw = self.vehicle.attitude.yaw
         except:
             pass
+
         #Treat batery apart, as it uses to fail
         try:
             self.status.battery = float(self.vehicle.battery.voltage)
         except:
             pass
+
         #finally if one of these values fails, dont send the message, this values are always reported
         try:
             self.status.armed = self.vehicle.armed
@@ -190,8 +292,6 @@ class Dronekit_node(Node):
             self.status_publisher.publish(self.status)
         except:
             pass
-        
-
 
     def get_bearing(self, location1, location2):
         """
@@ -212,44 +312,23 @@ class Dronekit_node(Node):
             bearing += 360.00
         return bearing
 
+    def call_service(self, client,  msg):
+        # TODO: raise error to avoid infinite wait if service is not up, after all means a module is not active $$ watchdog will be in charge
+        if not client.wait_for_service(timeout_sec=10.0):
+            self.get_logger().info(f'service {client.srv_name} not available', once=True)
+            return False
+        future = client.call_async(msg)
+        while rclpy.ok():
+            if future.done():
+                try:
+                    response = future.result()
+                except Exception as e:
+                    self.get_logger().info(
+                        'Service call failed %r' % (e,))
+                else:
+                    return response
+                break
 
-    def condition_yaw(self, heading, relative=False):
-        """
-        Send MAV_CMD_CONDITION_YAW message to point vehicle at a specified heading (in degrees).
-        This method sets an absolute heading by default, but you can set the `relative` parameter
-        to `True` to set yaw relative to the current yaw heading.
-        By default the yaw of the vehicle will follow the direction of travel. After setting
-        the yaw using this function there is no way to return to the default yaw "follow direction
-        of travel" behaviour (https://github.com/diydrones/ardupilot/issues/2427)
-        For more information see:
-        http://copter.ardupilot.com/wiki/common-mavlink-mission-command-messages-mav_cmd/#mav_cmd_condition_yaw
-        """
-        if relative:
-            is_relative = 1  # yaw relative to direction of travel
-        else:
-            is_relative = 0  # yaw is an absolute angle
-        # create the CONDITION_YAW command using command_long_encode()
-        msg = self.vehicle.message_factory.command_long_encode(
-            0, 0,  # target system, target component
-            pymavlink.mavutil.mavlink.MAV_CMD_CONDITION_YAW,  # command
-            0,  # confirmation
-            heading,  # param 1, yaw in degrees
-            0,  # param 2, yaw speed deg/s
-            1,  # param 3, direction -1 ccw, 1 cw
-            is_relative,  # param 4, relative offset 1, absolute angle 0
-            0, 0, 0)  # param 5 ~ 7 not used
-        # send command to vehicle
-        self.vehicle.send_mavlink(msg)
-
-    def reached_position(self, goal_loc):
-        """
-        Args:
-            goal_loc: Reference position (dronekit.LocationGlobal).
-        Returns:
-            'True' if the ASV distance respecto to the target Waypoint is less than 0.5 meters.
-        """
-
-        return self.calculate_distance(goal_loc) < 1.5
 
     def calculate_distance(self, goal_loc):
         """
@@ -279,373 +358,197 @@ class Dronekit_node(Node):
         return 6378100.0 * c
 
     def change_asv_mode_callback(self, request, response):
-        #string takes preference before int
+        """ Change the ASV mode """
+
         if self.status.manual_mode:
             self.get_logger().info(f"RC in manual, cannot change mode")
             return response
-        if len(request.asv_mode_str) != 0:
-            if request.asv_mode_str in self.mode_type.values():
-                mode=request.asv_mode_str
-            else:
-                self.get_logger().info(f"{request.asv_mode_str} is not a valid mode")
-                response.success=False
-                return response
+
+        # Check if the mode is valid
+        if request.asv_mode_str in ardupilotModeDictionary["ASV_MODE"].values():
+            mode = request.asv_mode_str
         else:
-            if request.asv_mode in self.mode_type:
-                mode=self.mode_type[str(request.asv_mode)]
-            else:
-                self.get_logger().info(f"{request.asv_mode} is not a valid mode")
-                response.success = False
-                return response
+            self.get_logger().info(f"{request.asv_mode_str} is not a valid mode")
+            response.success = False
+            return response
+
+        # Change the mode
         self.vehicle.mode = VehicleMode(mode)
-        response.success=True
+        response.success = True
         return response
-        #TODO: add other responses
+    
+    def goto_check_goal_callback(self, goal_request):
+        """ This callback is called when a new goal is received. 
+        It checks if the goal is valid """
 
-    ######################################### ACTION GOTO DESCRIPTION ############################################
+        self.get_logger().info("Received new point to go to")
 
-    def goto_accept(self, goal_request):
-        # This callback can optionally be implemented, and is called when a new goal is requested.
-        # Here we check if we can accept the goal or not, based on the current state of the system.
-
-        self.get_logger().info(f'Action call received')
-        #if we are attending another call exit
-        if self.goto_goal_handle is not None and self.goto_goal_handle.is_active:
-            self.get_logger().error(f'Action is busy')
-            return GoalResponse.REJECT
+        # Check if the goal is valid
         if not self.vehicle.armed or self.vehicle.mode != VehicleMode("LOITER"):
             self.get_logger().error(f'Error: vehicle should be armed and in loiter mode\nbut arming is {self.vehicle.armed} and vehicle is in {self.vehicle.mode}.')
             if not self.vehicle.ekf_ok:
                 self.get_logger().error(f"EKF seems to be the main issue, System Status: mode {self.vehicle.mode.name}, GPS_status: {self.vehicle.gps_0}, System status: {self.vehicle.system_status.state}, System able to arm {self.vehicle.is_armable}")
             return GoalResponse.REJECT #To move we must be armed and in loiter
+
         if self.status.manual_mode:
             self.get_logger().error(f'RC in manual mode, action rejected')
             return GoalResponse.REJECT
+
         if self.calculate_distance(goal_request.samplepoint)> self.max_distance:
             self.get_logger().error(f'we are too far away from the destination point {self.calculate_distance(goal_request.samplepoint)}m, cancelling action')
             return GoalResponse.REJECT
+
         self.get_logger().info(f'Action accepted')
         return GoalResponse.ACCEPT
 
+    def goto_cancel_callback(self,goal_handle):
+        """ This callback is called when the action is cancelled """
 
-
-    def goto_accepted_callback(self, goal_handle):
-        # This callback is called when the action server has accepted the client request to begin a new goal.
-        """
-        self.goto_goal_handle=goal_handle
-        if self.calculate_distance(goal_handle.request.samplepoint)> self.max_distance:
-            self.get_logger().warning(f'we are too far away from the destination point {self.calculate_distance(goal_handle.request.samplepoint)}m, cancelling action')
-            #result.finish_flag = "Point too far away"
-            goal_handle.abort()
-            return Goto.Result()
-        """
-        self.get_logger().info(f'Executing action: The mission has been accepted and is being processed')
-        goal_handle.execute() #execute goto_execute_callback
-
-    def goto_cancel(self, goal_handle):
-        """Accept or reject a client request to cancel an action."""
-        self.get_logger().info('Received cancel request')
+        self.get_logger().info("Action cancelled")
         return CancelResponse.ACCEPT
 
+    def goto_execute_callback(self, goal_handle): 
+        """ This callback is called when a new goal is received.
 
-    def goto_execute_callback(self, goal_handle): #TODO: check correct behaviour
+        It sends the vehicle to the given location.
+
+        """
+        # Feedback message is the remaining distance to the goal
         feedback_msg = Goto.Feedback()
-        result=Goto.Result()
-        arm_counter=0 #counter to leave if several seconds disarmed
-        ekf_counter=0 #counter to leave if several seconds with ekf in bad state
-        ekf_failed=False #bool to know were we come from
+        # Result message is a boolen indicating if the goal was reached
+        result = Goto.Result()
 
-        #Calculate Path
-        destination=Newpoint.Request()
-        destination.new_point=goal_handle.request.samplepoint
-        path=self.call_service(self.calculate_path_client, destination)
-        #if there is no path, go directly to destination
-        if path==False or path.success==False:
-            self.get_logger().error("something went wrong with path planner")
-            #result.finish_flag = "Planner couldn't calculate point"
-            #goal_handle.abort()
-            #return result
-            self.get_logger().info("going to point without planner")
-            path=[goal_handle.request.samplepoint] #TODO: decide whether to go directly to point or cancel action
+        # Compute the path #
+        destination_point = Newpoint.Request()
+        destination_point.new_point = goal_handle.request.samplepoint
+
+        # Compute the free path to the goal
+        path = self.call_service(self.calculate_path_client, destination_point)
+
+        # Check if the path is valid
+        if path == False or path.success == False:
+            self.get_logger().error("The planner could not find a valid path")
+            self.get_logger().error("Going to the goal directly without checking for obstacles")
+            # If the path is not valid, go directly to the goal
+            path = [goal_handle.request.samplepoint]
         else:
-            #extract path
-            path=path.point_list
+            # If the path is valid, go to the first waypoint
+            path = path.point_list
+
+        # Set the vehicle to GUIDED mode
         self.vehicle.mode = VehicleMode("GUIDED")
-        #self.condition_yaw(self.get_bearing(self.vehicle.location.global_relative_frame, goal_handle.request.samplepoint))
-        time.sleep(1)
-        while (rclpy.ok()) and (len(path)>0):
-            travelling_counter=0 #counter time since going to waypoint
-            self.vehicle.simple_goto(LocationGlobal(path[0].lat, path[0].lon, 0.0))
-            if len(path)>1:
-                location1= Location()
-                location1.lat=path[0].lat
-                location1.lon=path[0].lon
-                self.waypoints_publisher.publish(location1)
-            while rclpy.ok() and not self.reached_position(path[0]): #while we reach the goal, check for events
-                if goal_handle.is_cancel_requested:#event mission canceled
-                    #make it loiter around actual position
+
+        # Go to iteratively to each waypoint
+        arm_counter = 0
+
+        while rclpy.ok() and len(path) > 0:
+
+            # Get the next waypoint
+            next_waypoint = path.pop(0)
+            self.get_logger().info(f"Going to {next_waypoint}")
+
+            # Send the vehicle to the next waypoint
+            self.vehicle.simple_goto(LocationGlobal(next_waypoint.lat, next_waypoint.lon, 0.0))
+            self.publish_waypoint(next_waypoint)
+            
+            # Wait until the vehicle is close to the waypoint
+            while rclpy.ok() and not self.reached_position(next_waypoint):
+                # Publish the remaining distance to the goal
+                feedback_msg.distance = self.calculate_distance(next_waypoint)
+                goal_handle.publish_feedback(feedback_msg)
+                time.sleep(0.1)
+
+                # Check the abort conditions #
+                # 1) The action is cancelled -> To loiter around actual position #
+                if goal_handle.is_cancel_requested:
                     goal_handle.canceled()
                     self.vehicle.mode = VehicleMode("LOITER")
-                    self.get_logger().info('Goto Action canceled')
+                    self.get_logger().info("GOTO Action cancelled")
                     result.finish_flag = "Action cancelled"
                     return result
-                
+
+                # 2) The vehicle is in manual mode -> To loiter around actual position #
                 elif self.status.manual_mode:
-                        self.get_logger().warning("manual interruption during mission")
+                        self.get_logger().warning("Manual interruption during mission")
                         result.finish_flag = "Manual interruption"
                         goal_handle.abort()
                         return result
 
+                # 3) EKF is not ok 
                 elif not self.vehicle.ekf_ok: #if ekf is not ok, stop
-                    if self.vehicle.mode != VehicleMode("MANUAL"):
-                        self.get_logger().warning('EKF FAILED, sensor read not good enough, waiting for better signal, switching to manual and disarming')
-                        self.vehicle.mode = VehicleMode("MANUAL")
-                        self.vehicle.disarm()
-                        ekf_failed=True
-                    elif ekf_counter%30 == 0:
-                            self.get_logger().info(f"System Status: \nmode {self.vehicle.mode.name}, GPS_status: {self.vehicle.gps_0}, System status: {self.vehicle.system_status.state}, System able to arm {self.vehicle.is_armable} ")
-                    if ekf_counter > self.ekf_timeout*10:
-                        result.finish_flag = "ekf timeout"
-                        self.get_logger().info('EKF timeout, exiting action')
-                        goal_handle.abort()
-                        return result
-                    ekf_counter+=1
-                
+
+                    self.get_logger().warning('EKF FAILED, sensor read not good enough, waiting for better signal, switching to manual and disarming')
+                    self.vehicle.mode = VehicleMode("MANUAL")
+                    self.vehicle.disarm()
+                    result.finish_flag = "ekf timeout"
+                    self.get_logger().info('EKF timeout, exiting action')
+                    goal_handle.abort()
+                    return result
+
+                # 4) The vehicle is not in GUIDED mode -> vehicle is not in desired mode
                 elif self.vehicle.mode != VehicleMode("GUIDED"): #vehicle is not in desired mode
                     self.get_logger().warning("asv_mode was changed externally")
-                    if ekf_failed:
-                        self.get_logger().info("EKF failsafe cleared, resuming mission")
-                        self.get_logger().info(f"System Status: mode {self.vehicle.mode.name}, GPS_status: {self.vehicle.gps_0}, System status: {self.vehicle.system_status.state}, System able to arm {self.vehicle.is_armable} ")
-                        ekf_failed=False #reset flag
-                        ekf_counter=0
-                        self.vehicle.arm() #arm to avoid inconsistent state
-                    self.vehicle.mode = VehicleMode("GUIDED") #restore mode
-                    self.condition_yaw(self.get_bearing(self.vehicle.location.global_relative_frame, goal_handle.request.samplepoint))
+                    self.vehicle.mode = VehicleMode("GUIDED") # restore mode
                     self.vehicle.simple_goto(LocationGlobal(goal_handle.request.samplepoint.lat, goal_handle.request.samplepoint.lon, 0.0))
-            
+
+                # 5) The vehicle is not armed -> vehicle is not armed
                 elif not self.status.armed: #event vehicle disarmed
+
                     if not self.vehicle.is_armable:
                         self.get_logger().fatal('vehicle is not armable, inconsistent state')
+                    
                     else:
-                        if arm_counter%10==0:
+                        if arm_counter % 10 == 0:
                             self.get_logger().info('received external disarm in auto mode, waiting for RC')
-                        if self.vehicle.channels['5']>1200:
+                        
+                        # Wait for RC to arm
+                        if self.vehicle.channels['5'] > 1200:
                             self.vehicle.arm()
-                            self.get_logger().info('vehicle armed')
+                            self.get_logger().info('Vehicle armed!')
                             arm_counter=0
                         else:
                             arm_counter+=1
-                        if arm_counter > self.arm_timeout*10:
+
+                        # If the vehicle is not armed after arm_timeout seconds, abort the action
+                        if arm_counter > self.arm_timeout:
                             self.get_logger().info('Waited for RC for too long, exiting action')
                             result.finish_flag = "timeout vehicle disarmed"
                             goal_handle.abort()
                             return result
 
-                if travelling_counter>self.travelling_timeout*10:
-                    self.get_logger().info('timeout reaching the point, exiting action')
-                    result.finish_flag = "timeout reaching point"
-                    goal_handle.abort()
-                    return result
-                
-                if travelling_counter%30 == 0: #each 3 seconds
-                    feedback_msg.distance = self.calculate_distance(path[0])
-                    goal_handle.publish_feedback(feedback_msg)
+            # Goal reached
+            self.get_logger().info(f"Reached {next_waypoint}!")
 
-                travelling_counter+=1
-                time.sleep(0.1)
-            #waypoint reached so go to next one
-            if len(path)>1:
-                self.get_logger().info(f"waypoint reached going to [{path[1].lat},{path[1].lon}")
-            path.pop(0)
-        # after reaching samplepoint
-        self.vehicle.mode = VehicleMode("LOITER")
+        # If the vehicle is in GUIDED mode, set it to LOITER
+        if self.vehicle.mode == VehicleMode("GUIDED"):
+            self.vehicle.mode = VehicleMode("LOITER")
         goal_handle.succeed()
-        self.get_logger().info('Goal reached, waiting for sample')
+        self.get_logger().info("Final goal reached!")
 
-        if self.DEBUG:
-            self.waiting_for_sensor_read = False
-        else:
-            self.waiting_for_sensor_read = True
-
-        self.sensor_action_client.wait_for_server()
-        goal_msg = SensorSample.Goal()
-
-        self._send_goal_future = self.sensor_action_client.send_goal_async(goal_msg, feedback_callback=self.sensor_feedback_callback)
-        self._send_goal_future.add_done_callback(self.sensor_response)
-        while rclpy.ok() and self.waiting_for_sensor_read:
-            if goal_handle.is_cancel_requested:#event mission canceled
-                    goal_handle.canceled() #acknowledge goto cancelled
-                    self.get_logger().info('Goto Action canceled')
-                    result.finish_flag = "Action cancelled"
-                    self.cancel_sensor_read() #stop sensor read
-                    return result
-                
-            elif self.status.manual_mode:
-                    self.get_logger().warning("manual interruption during mission")
-                    result.finish_flag = "Manual interruption"
-                    self.cancel_sensor_read()
-                    goal_handle.abort()
-                    return result
-        #self.get_logger().debug(f'Sample value {value}')
         result.success = True
         result.finish_flag = "Normal execution"
         return result
 
-
-
-    ####################################### END ACTION DEFINITION ############################################
-
-
-
-    def call_service(self, client,  msg):
-        # TODO: raise error to avoid infinite wait if service is not up, after all means a module is not active $$ watchdog will be in charge
-        if not client.wait_for_service(timeout_sec=10.0):
-            self.get_logger().info(f'service {client.srv_name} not available', once=True)
-            return False
-        future = client.call_async(msg)
-        while rclpy.ok():
-            if future.done():
-                try:
-                    response = future.result()
-                except Exception as e:
-                    self.get_logger().info(
-                        'Service call failed %r' % (e,))
-                else:
-                    return response
-                break
-
-    # def rc_read_callback_debug(self):
-    #     self.get_logger().info("RC debug")
-
-    def rc_read_callback(self):
-        if self.DEBUG==False:
-            #if there is no RC return home
-            try:
-                #manage no RC detected
-                if all([self.vehicle.channels['5'] == 0, self.vehicle.channels['6'] == 0]):
-                    if self.vehicle.mode != VehicleMode("RTL"):
-                        self.get_logger().info("seems there is no RC connected", once=True)
-                        self.status.manual_mode = False #override RC if there is no connection
-                        #self.vehicle.mode = VehicleMode("RTL")
-                #manage RC switch between auto and manual mode
-                elif self.status.manual_mode!=(self.vehicle.channels['6']>1200):
-                    self.get_logger().info("manual interruption" if self.vehicle.channels['6']>1200 else "automatic control regained")
-                    self.status.manual_mode=bool(self.vehicle.channels['6']>1200)
-
-                #manage arm when RC in manual
-                elif self.status.manual_mode:                
-                    if self.vehicle.mode != VehicleMode("MANUAL") : #vehicle is not in desired mode
-                        self.get_logger().info("manual switching vehicle mode to manual")
-                        self.vehicle.mode = VehicleMode("MANUAL")
-                    if self.vehicle.armed!=(self.vehicle.channels['5']>1200):
-                        if self.vehicle.channels['5']>1200:
-                            self.vehicle.arm()
-                            self.get_logger().info("manual arm")
-                        else:
-                            self.vehicle.disarm()
-                            self.get_logger().info("manual disarm")
-            except:
-                pass
-
-    def vehicle_status_callback(self, vehicle, name, msg):
-        pass
-        #self.get_logger().info(f"{name} : {msg}")
-
-    def dictionary(self):
-        self.mode_type=dictionary("ASV_MODE")
-        
     
-    """
-        This function forces the stop of a sensor read
-    """   
-    def cancel_sensor_read(self):
-        if self.waiting_for_sensor_read: #if the action was active
-            self.get_logger().info("the asv was taking a sample, canceling")
-            self.sensor_goal_handle.cancel_goal_async()
-        pass
+    def reached_position(self, goal_loc):
+        """
+        Args:
+            goal_loc: Reference position (dronekit.LocationGlobal).
+        Returns:
+            'True' if the ASV distance respecto to the target Waypoint is less than 0.5 meters.
+        """
+
+        return self.calculate_distance(goal_loc) < self.wp_radius
+
+    def publish_waypoint(self, waypoint):
+        """ Publish the waypoint to the waypoint_mark topic """
+
+        waypoint_msg = Location()
+        waypoint_msg.lat = waypoint.lat
+        waypoint_msg.lon = waypoint.lon
+        self.waypoints_publisher.publish(waypoint_msg)
 
 
-    """
-        This function prints the feedback from the goto action
-    """     
-    def sensor_feedback_callback(self, feedback):
-        sensor_read= feedback.feedback.sensor_read #TODO: unused
-
-    """
-        This function is the first answer from the action service, indicates whether the request is accepted or rejected
-    """   
-    def sensor_response(self, future):
-        self.goal_handle = future.result()
-        if not self.goal_handle.accepted:
-            self.get_logger().error('Sensor read rejected :(, something is not working')
-            return
-        self.get_logger().debug('Sensor read accepted :)')
-        self._get_result_future = self.goal_handle.get_result_async()
-        self._get_result_future.add_done_callback(self.go_to_finished)
-
-
-    """
-        This function prints the result once the action finishes
-    """   
-    def sensor_read_finished(self, future):
-        result = future.result().result #this variable is the result specified in the action
-        status = future.result().status # ABORT = 4     CANCELED = 5    CANCEL_GOAL = 2     EXECUTE = 1     SUCCEED = 3     NOT_SPECIFIED = 6
-        status_string= ["0 IS NOT AN STATE", "EXECUTE", "CANCEL_GOAL", "SUCCEEDED", "ABORT", "CANCELED", "NOT SPECIFIED"]
-        if status == GoalStatus.STATUS_SUCCEEDED:
-            self.get_logger().debug('Goal succeeded! Result: {0}'.format(result.success))
-        else:
-            self.get_logger().info(f'Sensor read failed with status: {status_string[int(status)]}')
-            self.get_logger().info(f'reason : {result.finish_flag}')           
-        self.waiting_for_sensor_read = False # we indicate that sensor read has finished
-
-
-
-    #TODO: Dronekit is not able to send this message
-    """
-    This function is called as a callback from the topic obstacles
-    @params distance, angle_increment:   arrays indicating in each index distance and angle occupied by an obstacle.
-    if distance is greater than 20m then it is considered no obstacle. maximum ammount inpts 72.
-    
-    the data received should be a mean of the distance towards the object in order to achieve less than 72 obstacles in a 180º fov.
-    check https://ardupilot.org/dev/docs/code-overview-object-avoidance.html
-    """
-    def camera_obstacles_callback(self, msg):
-
-        angle_increment=msg.angle_increment
-
-        distance=[2000 for i in range(72)] #2000 or greater is no obstacle
-
-        if len(msg.distance)>72: #max size of array is 72
-            size=72
-        else:
-            size=len(msg.distance)
-
-        for i in range(size):
-            distance[i]=uint(msg.distance[i])
-
-
-        msg2 = mavlink2.MAVLink_obstacle_distance_message(
-            0,                      # us Timestamp (UNIX time or time since system boot)
-            3,                      # sensor_type, defined here: https://mavlink.io/en/messages/common.html#MAV_DISTANCE_SENSOR
-            distance,               # distances,    uint16_t[72],   cm
-            int(3),                 # increment,    uint8_t,        deg
-            int(30),	            # min_distance, uint16_t,       cm
-            int(2000),              # max_distance, uint16_t,       cm
-            float(angle_increment),	            # increment_f,  float,          deg #https://mavlink.io/en/messages/common.html#OBSTACLE_DISTANCE
-            float(-54), # angle_offset, float,          deg
-            12                      # MAV_FRAME, vehicle-front aligned: https://mavlink.io/en/messages/common.html#MAV_FRAME_BODY_FRD    
-            )
-
-        self.get_logger().info(f"{msg2}",once=True)
-        try:
-            self.vehicle.send_mavlink(msg2)
-        except:
-            self.get_logger().info(f'obstacle send failed:{traceback.format_exc()}')    
-
-
-    def reset_home_callback(self, request, response):
-        self.vehicle.home_location = self.vehicle.location.global_frame
-        
 def main(args=None):
     #init ROS2
     rclpy.init(args=args)
